Add handler to regenerate API Key button
