+++ client/app/pages/users/show.html
@@ -14,17 +14,6 @@
           <button type="button" class="btn btn-primary" ng-click="enableUser(user)">Enable</button>
         </div>
 
-        <div ng-if="!user.is_disabled">
-          <div class="form-group">
-            <label class="control-label">API Key</label>
-            <input type="text" class="form-control" value="{{user.api_key}}" size="44" readonly/>
-          </div>
-          <div class="form-group">
-            <button class="btn btn-default" ng-click="regenerateUserApiKey(user)" ng-disabled="disableRegenerateApiKeyButton">Regenerate</button>
-          </div>
-
-          <hr>
-
           <button type="button" class="btn btn-default" ng-click="isCollapsed = !isCollapsed">Change password <span class="caret"></span></button>
 
           <div uib-collapse="isCollapsed">
